Possible fix for internal chat names in push notifications
