+++ AdamantShared/Helpers/AdamantContacts.swift
@@ -1,0 +1,102 @@
+//
+//  AdamantContacts.swift
+//  Adamant
+//
+//  Created by Stanislav Jelezoglo on 22.12.2022.
+//  Copyright © 2022 Adamant. All rights reserved.
+//
+
+import Foundation
+
+enum AdamantContacts {
+    case adamantBountyWallet
+    case adamantNewBountyWallet
+    case adamantIco
+    case iosSupport
+    case adamantExchange
+    case betOnBitcoin
+    case donate
+    case adamantWelcomeWallet
+    
+    static let systemAddresses: [String] = {
+        return [AdamantContacts.adamantExchange.name, AdamantContacts.betOnBitcoin.name, AdamantContacts.adamantIco.name, AdamantContacts.adamantBountyWallet.name, AdamantContacts.adamantNewBountyWallet.name, AdamantContacts.donate.name, AdamantContacts.adamantWelcomeWallet.name]
+    }()
+    
+    var name: String {
+        switch self {
+        case .adamantWelcomeWallet:
+            return NSLocalizedString("Accounts.AdamantTokens", comment: "System accounts: ADAMANT Tokens")
+        case .adamantBountyWallet, .adamantNewBountyWallet:
+            return NSLocalizedString("Accounts.AdamantBounty", comment: "System accounts: ADAMANT Bounty")
+        case .adamantIco:
+            return "Adamant ICO"
+        case .iosSupport:
+            return NSLocalizedString("Accounts.iOSSupport", comment: "System accounts: ADAMANT iOS Support")
+        case .adamantExchange:
+            return NSLocalizedString("Accounts.AdamantExchange", comment: "System accounts: ADAMANT Exchange")
+        case .betOnBitcoin:
+            return NSLocalizedString("Accounts.BetOnBitcoin", comment: "System accounts: Bet on Bitcoin Price")
+        case .donate:
+            return NSLocalizedString("Accounts.DonateADMFoundation", comment: "System accounts: Donates ADAMANT Foundation")
+        }
+    }
+    
+    var isSystem: Bool {
+        switch self {
+        case .adamantExchange, .betOnBitcoin:
+            return false
+        default:
+            return true
+        }
+    }
+    
+    var address: String {
+        switch self {
+        case .adamantBountyWallet: return AdamantResources.contacts.adamantBountyWallet
+        case .adamantNewBountyWallet: return AdamantResources.contacts.adamantNewBountyWallet
+        case .adamantIco: return AdamantResources.contacts.adamantIco
+        case .iosSupport: return AdamantResources.contacts.iosSupport
+        case .adamantExchange: return AdamantResources.contacts.adamantExchange
+        case .betOnBitcoin: return AdamantResources.contacts.betOnBitcoin
+        case .donate: return AdamantResources.contacts.donateWallet
+        case .adamantWelcomeWallet: return AdamantResources.contacts.adamantWelcomeWallet
+        }
+    }
+    
+    var publicKey: String? {
+        switch self {
+        case .adamantExchange: return AdamantResources.contacts.adamantExchangePK
+        case .betOnBitcoin: return AdamantResources.contacts.betOnBitcoinPK
+        case .adamantBountyWallet: return AdamantResources.contacts.adamantBountyWalletPK
+        case .adamantNewBountyWallet: return AdamantResources.contacts.adamantNewBountyWalletPK
+        case .iosSupport: return AdamantResources.contacts.iosSupportPK
+        case .adamantIco: return AdamantResources.contacts.adamantIcoPK
+        case .donate: return AdamantResources.contacts.donateWalletPK
+        case .adamantWelcomeWallet: return AdamantResources.contacts.adamantBountyWalletPK
+        }
+    }
+    
+    var isReadonly: Bool {
+        switch self {
+        case .adamantBountyWallet, .adamantNewBountyWallet, .adamantIco, .adamantWelcomeWallet: return true
+        case .iosSupport, .adamantExchange, .betOnBitcoin, .donate: return false
+        }
+    }
+    
+    var isHidden: Bool {
+        switch self {
+        case .adamantBountyWallet, .adamantNewBountyWallet: return true
+        case .adamantIco, .iosSupport, .adamantExchange, .betOnBitcoin, .donate, .adamantWelcomeWallet: return false
+        }
+    }
+    
+    var avatar: String {
+        switch self {
+        case .adamantExchange, .betOnBitcoin, .donate, .adamantBountyWallet, .adamantNewBountyWallet:
+            return ""
+        default:
+            return "avatar_bots"
+        }
+    }
+    
+}
