+++ AdamantShared/Helpers/String+utilites.swift
@@ -38,3 +38,20 @@ extension String {
         return self[index(startIndex, offsetBy:i)]
     }
 }
+
+extension String {
+    func checkAndReplaceSystemWallets() -> String {
+        switch self {
+        case "chats.virtual.bounty_wallet_title":
+            return AdamantContacts.adamantNewBountyWallet.name
+        case "chats.virtual.bitcoin_bet_title":
+            return AdamantContacts.betOnBitcoin.name
+        case "chats.virtual.donate_bot_title":
+            return AdamantContacts.donate.name
+        case "chats.virtual.exchange_bot_title":
+            return AdamantContacts.adamantExchange.name
+        default:
+            return self
+        }
+    }
+}
